refactor: Extract EmojiCategories module from ConfigScreen.js (v1.0.51)
**Phase 4.5: UI Module Extraction - Step 5 of 9**

Extracted emoji categories section with complex master checkbox
synchronization logic to dedicated module.

**New File Created:**
- src/config/ui/EmojiCategories.js (267 lines)
  * render() - Returns emoji categories section HTML
  * attachEventListeners() - Wires category checkboxes and sliders
  * handleMasterToggle() - Syncs all categories when master changes
  * updateMasterCheckboxState() - Sets indeterminate state
  * updateVisualState() - CSS class management for enabled/disabled
  * loadConfig() - Loads emoji category config into UI
  * getConfigData() - Extracts emoji category config from UI

**Master Checkbox Logic:**
- All categories checked → master checked
- No categories checked → master unchecked
- Some categories checked → master indeterminate (half-ticked)

**Categories Managed:**
- Animals, Food, Vehicles, Faces, Nature, Objects
- Each with checkbox + weight slider

**Changes to ConfigScreen.js:**
- Added import for EmojiCategories module
- Replaced this.createEmojiCategoriesSection() call
- Removed 5 methods (179 lines total):
  * createEmojiCategoriesSection()
  * setupEmojiMasterToggle()
  * updateEmojiMasterCheckbox()
  * updateEmojiMasterCheckboxState()
  * updateEmojiCategoriesVisualState()
- Removed setupEmojiMasterToggle() call from addEventListeners

**Impact:**
- ConfigScreen.js: 1,911 → 1,732 lines (9.4% reduction)
- New module: 267 lines (complex state sync logic)
- Total progress: 3,284 → 1,732 lines (47% reduction from original)
- All 902 unit tests passing

**Testing:**
✅ All 902 tests passing
✅ Master checkbox sync logic preserved
✅ No functional changes

**Next:**
Phase 4.6: Extract AdvancedSection module (audio, grid, cleanup)
